+++ src/main/java/com/ipartek/formacion/dbms/dao/interfaces/ClienteDAO.java
@@ -17,6 +17,6 @@ public interface ClienteDAO extends DAOSetter {
 
 	public Cliente getByIdentificador(String identificador);
 
-	public Cliente getInforme(int codigo);
+	public Cliente getInforme(long codigo);
 
 }
